+++ src/index.ts
@@ -78,8 +78,9 @@ export type { StatusReporter } from "./lib/core/status-reporter";
 
 // Scripting
 export { sh, bash, nu, py, script, Script, fragment, Fragment, embedSh, rawScript, RawScript, unsafeAllowExit, languageFor, dedent, renderScript, Sh, Bash, Nushell, Python, EXIT_CODE_PATH } from "./lib/script";
+export { registerLanguage, unregisterLanguage, registeredLanguageNames, registeredExtensions, languageNameForExtension } from "./lib/script";
 export { scriptFromFile, lintCommandForFile, languageNameForFile } from "./lib/script/from-file";
-export type { ScriptLanguage, ScriptCtx, ScriptInput, ScriptObject, ScriptOptions, EmbedShOptions, LanguageName } from "./lib/script";
+export type { ScriptLanguage, ScriptCtx, ScriptInput, ScriptObject, ScriptOptions, EmbedShOptions, LanguageName, KnownLanguageName, ScriptTag, RegisterLanguageOptions } from "./lib/script";
 
 // Reporters
 export { GitHubStatusReporter, statusParam } from "./lib/reporters/github-status-reporter";